LayoutTests/imported/w3c/web-platform-tests/trusted-types/trusted-types-duplicate-names-list-report-only-expected.txt

CONSOLE MESSAGE: The Content Security Policy 'trusted-types a b c' was delivered in report-only mode, but does not specify a 'report-to'; the policy will have no effect. Please either add a 'report-to' directive, or deliver the policy via the 'Content-Security-Policy' header.
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'a' because it violates the following Content Security Policy directive: "trusted-types a b c"
CONSOLE MESSAGE: [Report Only] Refused to create a TrustedTypePolicy named 'd' because it violates the following Content Security Policy directive: "trusted-types a b c"

PASS TrustedTypePolicyFactory and policy list in CSP.
